Made it so that data is saved to the cache when sent to the user

+++ cache_allocator.h
@@ -3,9 +3,9 @@
 
 #include <stdint.h>
 
-char* alloc_mem(uint32_t size);
+char* alloc_mem(unsigned int size);
 void free_mem(char* mem);
-int init_alloc(int size);
+int init_alloc(size_t size);
 void finish_alloc();
 
 #endif
